requirements.txt
Flask==2.1.2
google-auth==1.35.0
google-auth-oauthlib==0.4.6
google-api-python-client==2.0.2
requests==2.26.0
